// src/components/features/HLSVideoPlayer/components/VideoPlayer/types.ts
export interface FullscreenApi {
  isIosNoSpec?: boolean
  requestFullscreen?: 'requestFullscreen'
  exitFullscreen?: 'exitFullscreen'
  fullscreenElement?: 'fullscreenElement'
  fullscreenEnabled?: 'fullscreenEnabled'
  fullscreenchange?: 'fullscreenchange'
  fullscreenerror?: 'fullscreenerror'
  fullscreen?: 'fullscreen'
}